[RestResourceAction] Fix a wrong type for receiving a return value of getResourceId()
RestResourceAction.cc:346:15: warning: comparison of constant 18446744073709551615 with expression of type 'uint32_t' (aka 'unsigned int') is always false
      [-Wtautological-constant-out-of-range-compare]
        if (actionId == INVALID_ID) {
            ~~~~~~~~ ^  ~~~~~~~~~~

+++ server/src/RestResourceAction.cc
@@ -339,7 +339,7 @@ void RestResourceAction::handlePut(void)
 	UnifiedDataStore *dataStore = UnifiedDataStore::getInstance();
 
 	ActionDef actionDef;
-	uint32_t actionId;
+	uint64_t actionId;
 
 	// action id
 	actionId = getResourceId();
